refactor(tests): add ComponentFixture tests
Remove old TestComponentBuilder tests, and keep relevant
ComponentFixture tests as component_fixture_spec.
